pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "proton-telegram-bot"
version = "2.0.0"
description = "Telegram bot for Proton equipment rental platform"
authors = [{name = "Proton Team"}]
dependencies = [
    "aiogram>=3.4.1",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "aiohttp>=3.9.5",
    "python-dotenv>=1.0.1",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
]

[project.scripts]
start-bot = "main:main"